Add hitpoints to actors, only large actors can crush other actors

+++ world/rules.cpp
@@ -20,6 +20,7 @@
 #include <world/world.hpp>
 #include <world/effects.hpp>
 #include <world/parts.hpp>
+#include <util/num.hpp>
 
 Actor get_player() {
   // TODO: Assert that the actor is registered in World.
@@ -27,6 +28,16 @@ Actor get_player() {
   // XXX: Fixed ID is problematic if we want to switch the player entity
   // around.
   return Actor(1);
+}
+
+/// Add results of four dice which can give -1, 0 or 1 with equal
+/// probabilities. Result is distributed in a crude approximation of a normal
+/// distribution.
+int fudge_roll() {
+  int result = 0;
+  for (int i = 0; i < 4; i++)
+    result += rand_int(3) - 1;
+  return result;
 }
 
 bool blocks_shot(Location location) {
@@ -39,10 +50,25 @@ bool blocks_sight(Location location) {
   return kind == wall_terrain || kind == void_terrain || kind == curtain_terrain;
 }
 
+bool can_crush(Actor actor, Actor crushee) {
+  return actor.as<Blob_Part>().big && !crushee.as<Blob_Part>().big;
+}
+
+bool blocks_movement(Actor actor) {
+  return actor.has<Blob_Part>();
+}
+
 bool action_walk(Actor actor, const Vec2i& dir) {
   auto loc = actor.location();
   auto new_loc = loc + dir;
   if (actor.can_pop(new_loc)) {
+    for (auto a : actors_on(actor.footprint(new_loc))) {
+      if (a == actor) continue;
+      // Uncrushable actors in the target area, abort movement.
+      if (blocks_movement(a) && !can_crush(actor, a))
+        return false;
+    }
+
     actor.push();
 
     // XXX Hacky. Player is tracked by the view space object.
@@ -50,8 +76,12 @@ bool action_walk(Actor actor, const Vec2i& dir) {
       move_view_pos(dir);
 
     for (auto a : actors_on(actor.footprint(new_loc))) {
-      msg("Crush!");
-      delete_actor(a);
+      if (blocks_movement(a)) {
+        // Crushing damages you.
+        damage(actor, a.as<Blob_Part>().armor / 2);
+        msg("Crush!");
+        delete_actor(a);
+      }
     }
     actor.pop(new_loc);
     // Energy cost for movement.
@@ -72,6 +102,7 @@ bool action_shoot(Actor actor, const Vec2i& dir) {
 
   for (loc = loc + dir; dist < range; loc = loc + dir) {
     dist++;
+
     bool hit_actor = false;
     for (auto& a : actors_at(loc)) {
       if (a != actor) {
@@ -82,8 +113,7 @@ bool action_shoot(Actor actor, const Vec2i& dir) {
 
     if (hit_actor) {
       msg("Zap!");
-      damage(loc);
-      explosion_fx(loc);
+      damage(loc, actor.as<Blob_Part>().damage);
       break;
     }
     if (blocks_shot(loc))
@@ -97,10 +127,19 @@ bool action_shoot(Actor actor, const Vec2i& dir) {
   blob.energy -= 100;
 }
 
-void damage(Location location) {
-  // TODO, lots more detail
-  for (auto a : actors_at(location)) {
-    delete_actor(a);
+void damage(Location location, int amount) {
+  for (auto a : actors_at(location))
+    damage(a, amount);
+}
+
+void damage(Actor actor, int amount) {
+  if (actor.has<Blob_Part>()) {
+    auto& blob = actor.as<Blob_Part>();
+    blob.armor -= amount;
+    if (blob.armor <= 0) {
+      explosion_fx(actor.location());
+      delete_actor(actor);
+    }
   }
 }
